% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.group.R
\name{generate.group}
\alias{generate.group}
\title{Generate simulated data}
\usage{
generate.group(
  n,
  J,
  k,
  model.size = 5,
  cortype = 1,
  rho = 0.5,
  sigma1 = 1,
  sigma2 = 1,
  seed = 1
)
}
\arguments{
\item{n}{The number of observations.}

\item{J}{The number of groups of interest.}

\item{k}{The group size of each group. Only even group structure is allowed here.}

\item{model.size}{The number of important groups in the underlying regression model.}

\item{cortype}{The correlation structure.
\code{cortype = 1} denotes the independence structure,
where the covariance matrix has \eqn{(i,j)} entry equals \eqn{I(i \neq j)}.
\code{cortype = 2} denotes the exponential structure,
where the covariance matrix has \eqn{(i,j)} entry equals \eqn{rho^{|i-j|}}.
code{cortype = 3} denotes the constant structure,
where the non-diagnoal entries of covariance
matrix are \eqn{rho} and diagonal entries are 1.}

\item{rho}{A parameter used to characterize the pairwise correlation in
predictors. Default is \code{0.5}..}

\item{sigma1}{The value controlling the strength of the coefficients. A large value implies large coefficients. Default \code{sigma2 = 1}.}

\item{seed}{random seed. Default: \code{seed = 1}.}
}
\value{
A \code{list} object comprising:
\item{x}{Design matrix of predictors.}
\item{y}{Response variable.}
\item{beta}{The coefficients used in the underlying regression model.}
\item{group}{The group index of each variable.}
\item{true.group}{The important groups in the group linear model.}
}
\description{
Generate simulated data for group linear model.
}
\examples{

# Generate simulated data
n <- 200
J <- 100
k <- 5
model.size <- 5
data <- generate.group(n, J, k, model.size)
str(data)
}
\author{
Yanhang Zhang
}
