% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplicing.R
\name{gsplicing}
\alias{gsplicing}
\title{Best Subset of Groups Selection in Linear Model}
\usage{
\method{gsplicing}{default}(
  x,
  y,
  group,
  method = c("sequential", "gsection"),
  C.max = 2,
  tune = c("GIC", "BIC", "AIC", "cv"),
  T.list,
  T.min = 1,
  T.max,
  pi,
  weight = rep(1, nrow(x)),
  max.iter = 10,
  nfolds = 5,
  ...
)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{n \times p}; each row is an observation
vector and each column is a predictor.}

\item{y}{The response variable of \code{n} observations.}

\item{group}{A vector of integers indicating the which group each variable is in.
For variables in the same group, they should be located in adjacent columns of \code{x}
and their corresponding index in \code{group} should be the same.
Denote the first group as \code{1}, the second \code{2}, etc.}

\item{method}{The method to be used to select the optimal model size. For
\code{method = "sequential"}, we solve the BSGS for each model size in \code{T.list}.
For \code{method = "gsection"}, we solve the BSGS with model size via a gold-section approach.}

\item{C.max}{The size of the exchange subset of groups. Default is \code{C.max = 2}.}

\item{tune}{The type of criterion for choosing the best model size.
Available options are \code{"GIC"}, \code{"BIC"}, \code{"AIC"} and \code{"cv"}.
Default is \code{tune = "GIC"}.}

\item{T.list}{An integer vector representing the alternative model sizes.
Only used for \code{method = "sequential"}. Default is \code{T.list = 1:min(N, round(n/(log(p)*min(table(group))))))}.}

\item{T.min}{The minimum value of model sizes. Only used for \code{method = "gsection"}. Default is \code{T.min = 1}.}

\item{T.max}{The maximum value of model sizes. Only used for \code{method = "gsection"}. Default is \code{T.max = min(N, round(n/(log(p)*min(table(group)))))}.}

\item{pi}{Threshold of the difference of the loss function in each iteration.}

\item{weight}{Observation weights. Default is \code{weight = 1} for each observation as default.}

\item{max.iter}{The maximum number of performing splicing algorithm.
In most of the case, only a few times of splicing iteration can guarantee the convergence.
Default is \code{max.iter = 10}.}

\item{nfolds}{The number of folds in cross-validation. Default is \code{nfolds = 5}.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{

}
\description{
An implementation of best subset of groups selection (BSGS) in linear model via the splicing approach.
}
\details{
Best subset of groups selection aims to  identify
a small subset of groups to achieve the best interpretability on the
response variable. We call this problem as Best Subset of Groups
Selection (BSGS), BSGS problem under the support size \eqn{T} is:
\deqn{min_{\beta}\frac{1}{2n}||y-X\beta||^2_2 s.t.||\beta||_{0,2}<= T,}
see Zhang(2021) for details. To find the optimal support size \eqn{T},
we provide various criterion like GIC, AIC, BIC and cross-validation to determine it.
}
\examples{
\donttest{
library(GSplicing)
n <- 200
J <- 100
k <- 5
model.size <- 5
data <- generate.group(n, J, k, model.size)
fit <- gsplicing(data$x, data$y, data$group)
all(fit$best_group == data$true.group)
}
}
\references{
Certifiably Polynomial Algorithm for Best Group Subset Selection. Zhang, Yanhang, Junxian Zhu, Jin Zhu, and Xueqin Wang (2021). arXiv preprint arXiv:2104.12576.

A polynomial algorithm for best-subset selection problem. Junxian Zhu, Canhong Wen, Jin Zhu, Heping Zhang, Xueqin Wang. Proceedings of the National Academy of Sciences Dec 2020, 117 (52) 33117-33123; DOI: 10.1073/pnas.2014241117
}
\author{
Yanhang Zhang, Yifan Wang, Xirui Zhao
}
